+++ lib/rubocop/cop/style/trailing_underscore_variable.rb
@@ -15,10 +15,13 @@ module RuboCop
       #   #good
       #   a, b, = foo()
       #   a, = foo()
+      #   *a, b, _ = foo()  => We need to know to not include 2 variables in a
+      #   a, *b, _ = foo()  => The correction `a, *b, = foo()` is a syntax error
       class TrailingUnderscoreVariable < Cop
         include SurroundingSpace
 
-        MSG = 'Do not use trailing `_`s in parallel assignment.'
+        MSG = 'Do not use trailing `_`s in parallel assignment.'.freeze
+        UNDERSCORE = '_'.freeze
 
         def on_masgn(node)
           left, = *node
@@ -60,11 +63,29 @@ module RuboCop
           first_offense = nil
 
           variables.reverse_each do |variable|
-            break unless variable.children.first == :_
+            var, = *variable
+            var, = *var
+            if allow_named_underscore_variables
+              break unless var == :_
+            else
+              break unless var.to_s.start_with?(UNDERSCORE)
+            end
             first_offense = variable
           end
 
+          return nil if first_offense.nil?
+
+          first_offense_index = variables.index(first_offense)
+          0.upto(first_offense_index - 1).each do |index|
+            return nil if variables[index].splat_type?
+          end
+
           first_offense
+        end
+
+        def allow_named_underscore_variables
+          @allow_named_underscore_variables ||=
+            cop_config['AllowNamedUnderscoreVariables']
         end
       end
     end